Refactor Main class into Calculator with added functionality
The Main class was renamed to Calculator and additional functionality was included. Some of these additions include the creation of basic calculator functions such as the ability to perform operations like addition, subtraction etc. In addition to the main class refactoring, a JTextField named 'display' was introduced to facilitate the display of user input and calculation results. A private method 'addButton' was also added to efficiently add buttons to the calculator interface. The code now illustrates a more functional and intuitive program resembling a calculator.